Can now select text file with the GUI.
Fixed starting folder for file selection.
Changed logic behind User class. 
Added comments 
Now use std::cerr for error instead of cout

+++ src/validate.cpp
@@ -3,6 +3,8 @@
 #include <cctype>
 #include <string>
 
+// Handles different validation aspects of the program
+
 bool isValidEmail(const std::string& userName)
 {
     std::regex pattern("^[A-Za-z]{2,}@[A-Za-z]+\\.[A-Za-z]{2,}"); // Ex: gg@gg.gg
@@ -19,7 +21,7 @@ bool equalsIgnoreCase(const std::string& a, const std::string& b)
 {
   if(a.size() != b.size())
     return false;
-  for (int i = 0; i < a.size(); i++)
+  for (size_t i = 0; i < a.size(); i++)
     if(std::tolower(a[i]) != std::tolower(b[i]))
       return false;
   return true;
